==> R_scripts/Observer-location/Step3_ModelEvaluation.Rmd <==
[Observer Centered BRT Model]
For Observer-location Oregon 2020 data
Goal: Evaluate species distribution models (CV statistics) for all species of birds from Oregon trained BRT

Load packages
```{r}
library(pROC)
library(foreach)
library(doParallel)
library(reticulate)
library(rgee)
library(ggplot2)
library(mice)
library(gbm)
library(terra)
library(raster)
library(dismo)
library(gbm)
library(tidyverse)
library(readr)
library(Metrics)
```

Load Test Data for observer-centered
```{r}
# Define the base directory
testdata_dir <- "/Volumes/T7/eBird_Oregon2020/distance/Distance-Sampling/BRT_train/Oregon2020/TestData/"

# Define file names
file_names <- c("ObsCentered_FixedRadius.csv", "ObsCentered_noRadius.csv", "ObsCentered_ESW.csv")

# Load and assign datasets to variables
fixed.test <- read.csv(paste0(testdata_dir, file_names[1])) # 100 m radius
noRadius.test <- read.csv(paste0(testdata_dir, file_names[2])) # 30 m radius
esw.test <- read.csv(paste0(testdata_dir, file_names[3])) # species-specific effective radius

head(fixed.test)
```

Load the Trained models for observer-centered strategy
```{r}
individual_species_oregon <- split(fixed.test, fixed.test$Common_Name)

model.dir.noradius <- "/Volumes/T7/eBird_Oregon2020/distance/Distance-Sampling/BRT_train/Oregon2020/TrainedModel/SurveyCentered_NoRadius"
noradius.models <- list.files(model.dir.noradius, pattern = "\\.rds$", full.names = TRUE, recursive = TRUE)

model.dir.fixedradius <- "/Volumes/T7/eBird_Oregon2020/distance/Distance-Sampling/BRT_train/Oregon2020/TrainedModel/SurveyCentered_FixedRadius"
fixedradius.models <- list.files(model.dir.fixedradius, pattern = "\\.rds$", full.names = TRUE, recursive = TRUE)

model.dir.esw <- "/Volumes/T7/eBird_Oregon2020/distance/Distance-Sampling/BRT_train/Oregon2020/TrainedModel/SurveyCentered_ESW"
esw.models <- list.files(model.dir.esw, pattern = "\\.rds$", full.names = TRUE, recursive = TRUE)



```

***Individual species deviance test***
```{r}
# Trained model
acron.noradius <- readRDS("/Volumes/T7/eBird_Oregon2020/distance/Distance-Sampling/BRT_train/Oregon2020/TrainedModel/SurveyCentered_NoRadius/Acorn Woodpecker.rds")

# Test dataset
noRadius.test


# Predict on test dataset
Acron.pred.noRadius <- predict.gbm(object = acron.noradius, 
                                         newdata = noRadius.test[noRadius.test$Common_Name == "Acorn Woodpecker", ],
                                         n.trees = acron.noradius$gbm.call$best.trees,
                                         type = "response")
length(Acron.pred.noRadius)
length(noRadius.test[noRadius.test$Common_Name == "Acorn Woodpecker", ]$Occur)


Acorn.comb <- data.frame(noRadius.test[noRadius.test$Common_Name == "Acorn Woodpecker", ]$Occur, Acron.pred.noRadius)

# Calculate deviance
Acron.pixel.dev <- calc.deviance(obs = noRadius.test[noRadius.test$Common_Name == "Acorn Woodpecker", ]$Occur, pred = Acron.pred.noRadius, calc.mean = TRUE, family = "poisson")


# Ensure observed and predicted values have the same length
if (length(Acron.pred.noRadius) != length(noRadius.test[noRadius.test$Common_Name == "Acorn Woodpecker", ]$Occur)) {
    stop("Length mismatch between predicted and observed values. Check your filtering or dataset.")
}

# Ensure no NA values exist in predictions or observations
if (any(is.na(Acron.pred.noRadius)) || any(is.na(noRadius.test[noRadius.test$Common_Name == "Acorn Woodpecker", ]$Occur))) {
    stop("NA values detected in observed or predicted data.")
}
```

***Individual species cross-validation CV statistics*** Acorn Woodpecker
```{r}
names(acron.noradius)

# Show cross-validation statistics
acron.noradius$cv.statistics

# Extract deviance mean & SE of deviance across cross-validation
  # Deviance.mean
acron.noradius$cv.statistics$deviance.mean
  # Deviance.se
acron.noradius$cv.statistics$deviance.se


```


***===Model Evaluation (CV statistic, Kendall)====***
We will use deviance to test difference of likelihood between the fitted model (BRTs) and the saturated model
- First Assign NA to NULL BRT models for each species & predictor extraction method
Load all models for a given species and calculate deviance values
```{r}
# Create empty dataframe to store evaluation results
result <- data.frame(Species = character(),
                     Extraction_Approach = character(),
                     Metric = character(),
                     Pixel_radius = numeric(),
                     Fixed_Radius = numeric(),
                     Effective_Radius = numeric())

load_models_for_species <- function(species_name) {
    
    # Initialize variables to store Kendall's rank coefficient values
    kendall_noRadius <- NA
    kendall_fixed <- NA
    kendall_esw <- NA
    
    # Initialize variables to store CV stats
    cv_meandeviance_noRadius <- NA
    cv_meandeviance_fixed <- NA
    cv_meandeviance_esw <- NA
    
    cv_SEdeviance_noRadius <- NA
    cv_SEdeviance_fixed <- NA
    cv_SEdeviance_esw <- NA

    
    #=== For 30 m pixel radius ===#
    noradius.perspecies <- grep(species_name, noradius.models, value = TRUE)
    
    if (length(noradius.perspecies) > 0) {
        noradius.perspecies.model <- readRDS(noradius.perspecies)
        
        if (!is.null(noradius.perspecies.model$gbm.call$best.trees)) {
            pred.noRadius <- predict.gbm(object = noradius.perspecies.model, 
                                         newdata = noRadius.test[noRadius.test$Common_Name == species_name, ],
                                         n.trees = noradius.perspecies.model$gbm.call$best.trees,
                                         type = "response")
            
            # Calculate Kendall's correlation
            kendall_noRadius <- cor(x = noRadius.test[noRadius.test$Common_Name == species_name, ]$Occur, 
                                    y = pred.noRadius, method = "kendall")
            
            # Get CV statistics from trained model itself
            cv_meandeviance_noRadius <- noradius.perspecies.model$cv.statistics$deviance.mean
            cv_SEdeviance_noRadius <- noradius.perspecies.model$cv.statistics$deviance.se
            
        }
    }

    #=== For 100 m fixed radius ===#
    fixedradius.perspecies <- grep(species_name, fixedradius.models, value = TRUE)
    
    if (length(fixedradius.perspecies) > 0) {
        fixedradius.perspecies.model <- readRDS(fixedradius.perspecies)
        
        if (!is.null(fixedradius.perspecies.model$gbm.call$best.trees)) {
            pred.fixed <- predict.gbm(object = fixedradius.perspecies.model, 
                                      newdata = fixed.test[fixed.test$Common_Name == species_name, ],
                                      n.trees = fixedradius.perspecies.model$gbm.call$best.trees,
                                      type = "response")

            # Calculate Kendall's correlation
            kendall_fixed <- cor(x = fixed.test[fixed.test$Common_Name == species_name, ]$Occur, 
                                 y = pred.fixed, method = "kendall")
            
            # Get CV statistics from trained model itself
            cv_meandeviance_fixed <- fixedradius.perspecies.model$cv.statistics$deviance.mean
            cv_SEdeviance_fixed <- fixedradius.perspecies.model$cv.statistics$deviance.se

        }
    }

    #=== For Species-specific effective radius ===#
    esw.perspecies <- grep(species_name, esw.models, value = TRUE)
    
    if (length(esw.perspecies) > 0) {
        esw.perspecies.model <- readRDS(esw.perspecies)
        
        if (!is.null(esw.perspecies.model$gbm.call$best.trees)) {
            pred.ESW <- predict.gbm(object = esw.perspecies.model, 
                                    newdata = esw.test[esw.test$Common_Name == species_name, ],
                                    n.trees = esw.perspecies.model$gbm.call$best.trees,
                                    type = "response")

            # Calculate Kendall's correlation
            kendall_esw <- cor(x = esw.test[esw.test$Common_Name == species_name, ]$Occur, 
                                 y = pred.ESW, method = "kendall")
            
            # Get CV statistics from trained model itself
            cv_meandeviance_esw <- esw.perspecies.model$cv.statistics$deviance.mean
            cv_SEdeviance_esw <- esw.perspecies.model$cv.statistics$deviance.se

        }
    }

    # Store results in a dataframe
      # Kendall rank correlation test
    per_bird.result.kendall <- data.frame(Species = species_name,
                                      Extraction_Approach = "Observer-location Approach",
                                      Metric = "Kendall",
                                      Pixel_radius = kendall_noRadius,
                                      Fixed_Radius = kendall_fixed,
                                      Effective_Radius = kendall_esw)
      # CV statistics
        # Mean deviance
    per_bird.result.CVdeviance <- data.frame(Species = species_name,
                                             Extraction_Approach = "Observer-location Approach",
                                             Metric = "CV_MeanDeviance",
                                             Pixel_radius = cv_meandeviance_noRadius,
                                             Fixed_Radius = cv_meandeviance_fixed,
                                             Effective_Radius = cv_meandeviance_esw)
        # SE deviance
    per_bird.result.CV.SEdeviance <- data.frame(Species = species_name,
                                                Extraction_Approach = "Observer-location Approach",
                                                Metric = "CV_SE_Deviance",
                                                Pixel_radius = cv_SEdeviance_noRadius,
                                                Fixed_Radius = cv_SEdeviance_fixed,
                                                Effective_Radius = cv_SEdeviance_esw)
    
    # Concatenate per bird results in one data frame
    per_bird.result <- rbind(per_bird.result.kendall,
                             per_bird.result.CVdeviance, per_bird.result.CV.SEdeviance)
    
    return(per_bird.result)
}


# Loop over each bird species name and load models
for (bird in names(individual_species_oregon)) {
  print(bird)
  per_bird.result <- load_models_for_species(bird)
  result <- rbind(result, per_bird.result)
}

```

Export as CSV
```{r}
write.table(result, file = "/Volumes/T7/eBird_Oregon2020/distance/Distance-Sampling/Evaluation/Oregon2020/evalMetric-ObserverLocationOR2020.csv", sep = ",", row.names = FALSE, fileEncoding = "UTF-8")

```
